+++ rule/CashTracker/Resources/scripts/cashtracker/rest/TagService.java
@@ -16,6 +16,7 @@ import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 
 import cashtracker.dao.TagDAO;
+import cashtracker.model.Errors;
 import cashtracker.model.Tag;
 import cashtracker.validation.TagValidator;
 import cashtracker.validation.ValidationError;
@@ -43,7 +44,6 @@ public class TagService extends RestProvider {
 	public Response get(@PathParam("id") long id) {
 		TagDAO dao = new TagDAO(getSession());
 		IAppEntity m = dao.findById(id);
-		System.out.println(((Tag) m).getName());
 		return Response.ok(m).build();
 	}
 
@@ -53,7 +53,7 @@ public class TagService extends RestProvider {
 	public Response create(Tag m) {
 		ValidationError ve = validator.validate(m);
 		if (ve.hasError()) {
-			return Response.ok(ve).status(Status.BAD_REQUEST).build();
+			return Response.status(Status.BAD_REQUEST).entity(ve).build();
 		}
 
 		TagDAO dao = new TagDAO(getSession());
@@ -69,7 +69,7 @@ public class TagService extends RestProvider {
 
 		ValidationError ve = validator.validate(m);
 		if (ve.hasError()) {
-			return Response.ok(ve).status(Status.BAD_REQUEST).build();
+			return Response.status(Status.BAD_REQUEST).entity(ve).build();
 		}
 
 		TagDAO dao = new TagDAO(getSession());
@@ -78,21 +78,28 @@ public class TagService extends RestProvider {
 
 	@DELETE
 	@Path("/{id}")
+	@Produces(MediaType.APPLICATION_JSON)
 	public Response delete(@PathParam("id") long id) {
 		TagDAO dao = new TagDAO(getSession());
 		Tag m = dao.findById(id);
 		if (m != null) {
-			dao.delete(m);
+			if (dao.existsTransactionByTag(m)) {
+				Errors msg = new Errors();
+				msg.setMessage("used");
+				return Response.status(Status.BAD_REQUEST).entity(msg).build();
+			} else {
+				dao.delete(m);
+			}
 		}
 		return Response.status(Status.NO_CONTENT).build();
 	}
 
 	@JsonRootName("tags")
-	class Tags extends ArrayList <IAppEntity> {
+	class Tags extends ArrayList <Tag> {
 
 		private static final long serialVersionUID = 1L;
 
-		public Tags(Collection <? extends IAppEntity> m) {
+		public Tags(Collection <? extends Tag> m) {
 			addAll(m);
 		}
 	}
